requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
chromadb>=0.4.22
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1
python-multipart>=0.0.6
